Moved `leftRotate` and `rightRotate` to their own file
